pyproject.toml
[project]
name = "proxmox-config"
version = "0.1.0"
description = "Proxmox VE configuration tool for home lab infrastructure"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "proxmoxer>=2.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "InquirerPy>=0.3.4",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0",
    "paramiko>=3.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
proxmox-config = "src.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
